organizes files into folders for maintainability
