--- server/_jest-setup_/global-setup.js
@@ -1,5 +1,0 @@
-const { start } = require("./mssql-container-setup");
-
-module.exports = async () => {
-  await start();
-};
