+++ server/_jest-setup_/global-setup/global-setup.js
@@ -1,0 +1,5 @@
+const { start } = require("../utils/mssql-container-setup");
+
+module.exports = async () => {
+  await start();
+};
