+++ server/_jest-setup_/global-setup/global-teardown.js
@@ -1,0 +1,5 @@
+const { stop } = require("../utils/mssql-container-setup");
+
+module.exports = async () => {
+  await stop();
+};
