--- server/_jest-setup_/global-teardown.js
@@ -1,5 +1,0 @@
-const { stop } = require("./mssql-container-setup");
-
-module.exports = async () => {
-  await stop();
-};
