+++ server/_jest-setup_/utils/server-setup.js
@@ -1,4 +1,4 @@
-const app = require("../server");
+const app = require("../../server");
 
 require('dotenv').config();
