+++ server/jest.config.js
@@ -59,10 +59,10 @@ module.exports = {
   // forceCoverageMatch: [],
 
   // A path to a module which exports an async function that is triggered once before all test suites
-  globalSetup: "./_jest-setup_/global-setup.js",
+  globalSetup: "./_jest-setup_/global-setup/global-setup.js",
 
   // A path to a module which exports an async function that is triggered once after all test suites
-  globalTeardown: "./_jest-setup_/global-teardown.js",
+  globalTeardown: "./_jest-setup_/global-setup/global-teardown.js",
 
   // A set of global variables that need to be available in all test environments
   // globals: {},
@@ -135,7 +135,7 @@ module.exports = {
   // setupFiles: [],
 
   // A list of paths to modules that run some code to configure or set up the testing framework before each test
-  setupFilesAfterEnv: ["./_jest-setup_/setup-after-env.js"],
+  setupFilesAfterEnv: ["./_jest-setup_/local-setup/setup-after-env.js"],
 
   // The number of seconds after which a test is considered as slow and reported as such in the results.
   // slowTestThreshold: 5,
